Add files via upload
Draft reconstruction of MTB H37Rv with the use of CarveMe. A gapfilling step was added to allow the model to grow on different known media. Minimum growth rate parameter was changed to 0.001 mmol/gDW/h and the minimum flux allowed in reactions was changed to 1e-6 to allow the model to incorporate reactions.